vese-projects-code/api/requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.0.11
MarkupSafe==2.1.1
python-dotenv==0.21.0
Werkzeug==2.2.2
